ros2gams - new config file format working

+++ tests/capnfiles/Header.capnp
@@ -3,5 +3,5 @@
 struct Header{
   seq @0 :UInt32;
   frameid @1: Text;
-  stamp @2: UInt32;
+  stamp @2: Float64;
 }
